refactor(cli-app): extract Maestro argv/env injection into maestro-inject.js
Move maybeInjectMaestroServer + maybeInjectScreenshotDir (and the private argv
helpers findTestSubcommandIdx / hasExistingPercyServerFlag / findTestOutputDirValue
+ constants) out of exec.js into a dedicated maestro-inject.js. exec.js returns
to a thin command (211 → 47 lines) whose app:exec callback calls the injectors;
index.js re-exports them from the new module. No behavior change — exec.test.js
43/43 unchanged (imports via the @percy/cli-app barrel).
