+++ packages/cli-app/src/index.js
@@ -1,2 +1,3 @@
 export { default, app } from './app.js';
-export { exec, start, stop, ping, maybeInjectMaestroServer, maybeInjectScreenshotDir } from './exec.js';
+export { exec, start, stop, ping } from './exec.js';
+export { maybeInjectMaestroServer, maybeInjectScreenshotDir } from './maestro-inject.js';
